smart-presentation-generator-backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-pptx==0.6.21
python-dotenv==0.19.1
google-generativeai>=0.3.2